+++ MiDAS/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.0.9042 - 17/04/20
++ removes unused pvalue_cutoff argument from runMiDAS function
++ fixes bug with conditional testing of glm models
+
+## 0.0.9041 - 9/04/20
++ fixes problem with placeholder (dummy variable) being constant. This caused coxme function to throw errors upon model definition.
+
+## 0.0.9040 - 16/02/20
++ add functionality for Grantham distance calcualtion between amino acid sequences (`distGrantham` function) and between HLA alleles (`hlaCallsGranthamDistance`).
++ add Grantham distances calculation for Class I HLA alleles to `prepareMiDAS` and `runMiDAS` functionality.
+
 ## 0.0.9039 - 18/12/19
 + now using interactions terms is possible through placeholder argument in runMiDAS and presence of dummy var in input data frame